refactor: separation between dev and prod envs; improved utilities; relocate e2e tests
