+++ .github/workflows/test-ci-dev.yaml
@@ -26,7 +26,7 @@ jobs:
             - name: Get Library Versions For Binary Caching
               id: cache-settings
               run: |
-                echo "FIREBASE_TOOLS=$(yarn list -s --depth=0 --pattern firebase-tools | tail -n 1 | sed 's/.*@//g')" >> $GITHUB_OUTPUT
+                  echo "FIREBASE_TOOLS=$(yarn list -s --depth=0 --pattern firebase-tools | tail -n 1 | sed 's/.*@//g')" >> $GITHUB_OUTPUT
 
             - name: Cache Firebase Emulator Binaries
               uses: actions/cache@v3
