+++ packages/actions/.env.test.default
@@ -16,4 +16,7 @@ CONFIG_PRESIGNED_URL_EXPIRATION_IN_SECONDS=
 AUTH_GITHUB_CLIENT_ID=
 AUTH_USER_EMAIL=
 AUTH_GITHUB_USER_PW=
-AUTH_GMAIL_USER_PW=
+AUTH_GMAIL_CLIENT_ID=
+AUTH_GMAIL_CLIENT_SECRET=
+AUTH_GMAIL_REDIRECT_URL=
+AUTH_GMAIL_REFRESH_TOKEN=
