+++ packages/backend/firestore.indexes.json
@@ -1,33 +1,33 @@
 {
-  "indexes": [
-    {
-      "collectionGroup": "ceremonies",
-      "queryScope": "COLLECTION",
-      "fields": [
+    "indexes": [
         {
-          "fieldPath": "state",
-          "order": "ASCENDING"
+            "collectionGroup": "ceremonies",
+            "queryScope": "COLLECTION",
+            "fields": [
+                {
+                    "fieldPath": "state",
+                    "order": "ASCENDING"
+                },
+                {
+                    "fieldPath": "endDate",
+                    "order": "ASCENDING"
+                }
+            ]
         },
         {
-          "fieldPath": "endDate",
-          "order": "ASCENDING"
+            "collectionGroup": "ceremonies",
+            "queryScope": "COLLECTION",
+            "fields": [
+                {
+                    "fieldPath": "state",
+                    "order": "ASCENDING"
+                },
+                {
+                    "fieldPath": "startDate",
+                    "order": "ASCENDING"
+                }
+            ]
         }
-      ]
-    },
-    {
-      "collectionGroup": "ceremonies",
-      "queryScope": "COLLECTION",
-      "fields": [
-        {
-          "fieldPath": "state",
-          "order": "ASCENDING"
-        },
-        {
-          "fieldPath": "startDate",
-          "order": "ASCENDING"
-        }
-      ]
-    }
-  ],
-  "fieldOverrides": []
+    ],
+    "fieldOverrides": []
 }
